Make Entry.Size and ChunkUploadResponse.Offset be int64 - fixes #19
This allows the code to deal with files > 2 GB in size.  Previous to
this change a file > 2 GB would cause a JSON unmarshal error.

This is a breaking change to the API.  However the Go compiler will
throw errors at the right places when compiling so should be easy to
fix up the code.

+++ crypto.go
@@ -199,7 +199,7 @@ func (db *Dropbox) DownloadAES(key []byte, src, rev string, offset int) (io.Read
 	var size int64
 	var err error
 
-	if in, size, err = db.Download(src, rev, offset); err != nil {
+	if in, size, err = db.Download(src, rev, int64(offset)); err != nil {
 		return nil, err
 	}
 	return NewAESDecrypterReader(key, in, int(size))
